Refactor filesize  components.
